moved fixtures to fixtures.py
as in conftest we have only fixtures, we created a fixture file to
organize them
